# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  HTTP400BadRequest:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  HTTP404NotFound:
    properties:
      code:
        example: 404
        type: integer
      message:
        example: Not Found
        type: string
    type: object
  HTTP500InternalServerError:
    properties:
      code:
        example: 500
        type: integer
      message:
        example: InternalServerError
        type: string
    type: object
  ImageMirror:
    properties:
      image:
        type: string
      items:
        items:
          $ref: '#/definitions/MirrorImageItem'
        type: array
      namespace:
        type: string
      pool:
        type: string
    type: object
...
  Message:
    properties:
      message:
        example: message
        type: string
    type: object
  MirrorImage:
    properties:
      image:
        type: string
      items:
        items:
          $ref: '#/definitions/MirrorImageItem'
        type: array
      namespace:
        type: string
      pool:
        type: string
    type: object
...
  MirrorList:
    properties:
      Local:
        items:
          $ref: '#/definitions/MirrorImage'
        type: array
      Remote:
        items:
          $ref: '#/definitions/MirrorImage'
        type: array
    type: object
  MirrorSetup:
    properties:
      host:
        type: string
      localClusterName:
        description: 미러링 상태
        type: string
      localToken:
        type: string
      mirrorPool:
        type: string
      privateKeyFile: {}
      remoteClusterName:
        description: 미러링 상태
        type: string
      remoteToken:
        type: string
    type: object
  MirrorStatus:
    properties:
      daemon_health:
        description: 미러링 데몬 상태
        type: string
      health:
        description: 미러링 상태
        type: string
      image_health:
        description: 이미지 상태
        type: string
      states:
        description: 이미지 상세
    type: object
...
info:
  contact:
    email: support@ablecloud.io
    name: 윤여천
    url: http://www.ablecloud.io
  description: This is a GlueAPI server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Glue-API
  version: v1.0
paths:
...
  /api/v1/mirror:
    delete:
      consumes:
      - multipart/form-data
      description: Glue 의 미러링 클러스터를 제거합니다.
      parameters:
      - description: Remote Cluster Host Address
        in: formData
        name: host
        required: true
        type: string
      - description: Remote Cluster PrivateKey
        in: formData
        name: privateKeyFile
        required: true
        type: file
      - description: Pool Name for Mirroring
        in: formData
        name: mirrorPool
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/MirrorSetup'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTP404NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: Delete Mirroring Cluster
      tags:
      - Mirror
    get:
      consumes:
      - application/x-www-form-urlencoded
      description: Glue 의 미러링 상태를 보여줍니다.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/MirrorStatus'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTP404NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: Show Status of Mirror
      tags:
      - Mirror
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Glue 의 미러링 클러스터를 구성합니다.
      parameters:
      - description: Local Cluster Name
        in: formData
        name: localClusterName
        required: true
        type: string
      - description: Remote Cluster Name
        in: formData
        name: remoteClusterName
        required: true
        type: string
      - description: Remote Cluster Host Address
        in: formData
        name: host
        required: true
        type: string
      - description: Remote Cluster PrivateKey
        in: formData
        name: privateKeyFile
        required: true
        type: file
      - description: Pool Name for Mirroring
        in: formData
        name: mirrorPool
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/MirrorSetup'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTP404NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary:  Setup Mirroring Cluster
      tags:
      - Mirror
...
  /api/v1/mirror/image:
    get:
      consumes:
      - application/x-www-form-urlencoded
      description: 미러링중인 이미지의 목록과 상태를 보여줍니다.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/MirrorList'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTP404NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: Show List of Mirrored Snapshot
      tags:
      - Mirror
  /api/v1/mirror/image/{mirrorPool}/{imageName}:
    delete:
      consumes:
      - application/x-www-form-urlencoded
      description: 이미지의 미러링을 비활성화 합니다.
      parameters:
      - description: pool
        in: path
        name: mirrorPool
        required: true
        type: string
      - description: imageName
        in: path
        name: imageName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTP404NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: Delete Mirrored Snapshot
      tags:
      - Mirror
    get:
      consumes:
      - application/x-www-form-urlencoded
      description: 미러링중인 이미지의 정보를 보여줍니다.
      parameters:
      - description: mirrorPool
        in: path
        name: mirrorPool
        required: true
        type: string
      - description: imageName
        in: path
        name: imageName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ImageMirror'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTP404NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: Show Infomation of Mirrored Snapshot
      tags:
      - Mirror
    patch:
      consumes:
      - application/x-www-form-urlencoded
      description: Glue 의 이미지에 미러링의 설정을 변경합니다.
      parameters:
      - description: Pool Name for Mirroring
        in: path
        name: mirrorPool
        required: true
        type: string
      - description: Image Name for Mirroring
        in: path
        name: imageName
        required: true
        type: string
      - description: Interval of image snapshot
        in: formData
        name: interval
        required: true
        type: string
      - description: Starttime of image snapshot
        in: formData
        name: startTime
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ImageMirror'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTP404NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: Patch Image Mirroring
      tags:
      - Mirror
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Glue 의 이미지에 미러링을 설정합니다.
      parameters:
      - description: Pool Name for Mirroring
        in: path
        name: mirrorPool
        required: true
        type: string
      - description: Image Name for Mirroring
        in: path
        name: imageName
        required: true
        type: string
      - description: Interval of image snapshot
        in: formData
        name: interval
        required: true
        type: string
      - description: StartTime of image snapshot
        in: formData
        name: startTime
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ImageMirror'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTP404NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: Setup Image Mirroring
      tags:
      - Mirror
  /api/v1/mirror/image/status/{mirrorPool}/{imageName}:
    get:
      consumes:
      - application/x-www-form-urlencoded
      description: Glue 의 이미지에 미러링상태를 확인합니다.
      parameters:
      - description: Pool Name for Mirroring
        in: path
        name: mirrorPool
        required: true
        type: string
      - description: Image Name for Mirroring
        in: path
        name: imageName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ImageStatus'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTP404NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: Show Mirroring Image Status
      tags:
      - Mirror
  /api/v1/mirror/image/promote/{mirrorPool}/{imageName}:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Glue 의 이미지를 Promote 합니다.
      parameters:
      - description: Pool Name for Mirroring
        in: path
        name: mirrorPool
        required: true
        type: string
      - description: Image Name for Mirroring
        in: path
        name: imageName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ImageStatus'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/HTTP400BadRequest'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/HTTP404NotFound'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/HTTP500InternalServerError'
      summary: Promote Image Mirroring
      tags:
      - Mirror
...
securityDefinitions:
  ApiKeyAuth:
    description: Description for what is this security definition being used
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"